requirements.txt
gradio>=4.0.0
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.0.0
llama-index>=0.10.0 
llama-index-llms-openai-like>=0.3.4